transferred get_example_data functions from OLD_database to database.py
