+++ scopesim/server/__init__.py
@@ -1,4 +1,4 @@
-from .database import (download_package,
+from .database import (download_packages,
                        list_packages,
                        download_example_data,
                        list_example_data)
